requirements.txt
# requirements.txt
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
flask-wtf==1.1.1
flask_sqlalchemy==3.0.4
psycopg2-binary==2.9.6
apscheduler==3.10.1
# Packages de test
pytest==7.4.0
behave==1.2.6